requirements.txt
# Core bot dependencies
python-telegram-bot==20.7
ollama==0.3.3

# Document processing
PyMuPDF==1.23.24
Pillow==10.1.0
python-docx==1.1.0
openpyxl==3.1.2

# OCR capabilities
easyocr==1.7.0
pytesseract==0.3.10

# Text processing and NLP
spacy==3.7.2
regex==2023.10.3

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Development
black==23.11.0
pytest==7.4.3
pytest-asyncio==0.21.1
